Working Bootloader runs on TB3CMB
V02.01.13 -- 08/24/2017
We have a working bootloader here I will tweak the design so that main
program code loads into FLASH sectors FLASHC through FLASHJ, allowing it
to use all available flash, asside from what is reserved for the
bootloader. This change is coordinated with  a change in the main
program, and works with I_001_F281xTB3CMB_012. The substance of this
change is a tweak to MCS.C to  recognize that a downloaded program with
start load address 0x3D8000 (start of FLASHJ) is the main program, and
loads in FLASHC through FLASHJ.  Also tweak the bl_jumpToMain( )
function to use 0x3D8000 as the vector for jumping to the main program,
and 0x3D8002 as a 2-word key field with a known value, to  assure that
the main program has been burned into FLASH